requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
django-filter==24.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
